requirements.txt
python-binance
pyTelegramBotAPI
python-dotenv
asyncio